backend/requirements.txt
# ===========================================
# MiroFish Backend Dependencies
# ===========================================
# Python 3.11+ required
# Install: pip install -r requirements.txt
# ===========================================

# ============= Khung lõi =============
flask>=3.0.0
flask-cors>=6.0.0

# ============= Liên quan đến LLM =============
# OpenAI SDK (thống nhất dùng định dạng OpenAI để gọi LLM)
openai>=1.0.0

# ============= Zep Cloud =============
zep-cloud==3.13.0

# ============= Mô phỏng mạng xã hội OASIS =============
# Khung mô phỏng mạng xã hội OASIS
camel-oasis==0.2.5
camel-ai==0.2.78

# ============= Xử lý tệp =============
PyMuPDF>=1.24.0
# Phát hiện mã hóa (hỗ trợ tệp văn bản không dùng UTF-8)
charset-normalizer>=3.0.0
chardet>=5.0.0

# ============= Thư viện tiện ích =============
# Nạp biến môi trường
python-dotenv>=1.0.0

# Xác thực dữ liệu
pydantic>=2.0.0
